Добавил логику и эндпоинт запуска кредита и внесения платежей. Рефакторинг. Тесты

+++ loans-service/src/main/java/com/synloans/loans/model/entity/loan/Loan.java
@@ -20,7 +20,7 @@ public class Loan {
     private Long id;
 
     @Column(name = "sum")
-    private Double sum;
+    private Long sum;
 
     @Column(name = "rate")
     private Double rate;
